Fix int32 overflow in qdq_mxfp4 for tensors above INT32_MAX elements
`launch_qdq_mxfp4` truncated `a.numel()` into an `int`, and
`qdq_mxfp4_kernel` indexed elements with an `int`. For a tensor with more
than INT32_MAX elements this wraps negative, still passes the
divisibility check, and launches with a garbage grid dimension, so the
process dies with "Memory access fault by GPU node-N ... Reason: Unknown".

Use 64-bit types for the element count and the element index, and check
the grid dimension explicitly. This mirrors what the sibling
mxfp4/dequantize.cu already does.

Minimal reproducer (bf16, hidden 14336) on MI300A / ROCm 7.2:

  2,132,795,392 elements (under INT32_MAX) -> OK
  2,162,155,520 elements (over  INT32_MAX) -> memory access fault

With this change both cases pass.

+++ quark/torch/kernel/hw_emulation/csrc/mxfp4/fake_kernels.cuh
@@ -152,7 +152,7 @@ template <
 __global__ void qdq_mxfp4_kernel(float_type* inp, float_type* out) {
   // Each thread handles one element.
 
-  int idx = blockIdx.x * blockDim.x + threadIdx.x;
+  int64_t idx = static_cast<int64_t>(blockIdx.x) * blockDim.x + threadIdx.x;
 
   float_type elem = inp[idx];
   float_type block_max = habs_impl(elem);
